Aplica filtro de notificação na Home e ignora filtro ao deletar todas
